update firestore mocks
